Modified func add_mean_spacing_and_others_pos in utilities.py
Removed the adding of mean spacings, it will be done one we know the number of nearest neighbors to consider (we don't know it at this stage).
Also, added the distance after the x and y position of each pedestrian, so this data is already in the extended dataframe. Furthermore, the others' positions are sorted by distance from the current pedestrian